Refactor API keys routes to keeperhub directory
Move implementation from app/api/keeperhub/keys/ to keeperhub/api/keys/
following the KeeperHub code organization pattern. The app/ route files
are now thin wrappers that re-export from the keeperhub/ implementation.
